Manejo de excepciones

+++ src/main/java/clinica/proyectoclinicaodontologica/service/OdontologoService.java
@@ -1,12 +1,14 @@
 package clinica.proyectoclinicaodontologica.service;
 
 
+import clinica.proyectoclinicaodontologica.exceptions.ResourceNotFoundException;
 import clinica.proyectoclinicaodontologica.model.Odontologo;
 import clinica.proyectoclinicaodontologica.repository.OdontologoRepository;
 import org.springframework.stereotype.Service;
 
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class OdontologoService {
@@ -27,15 +29,27 @@ public class OdontologoService {
     }
 
     public Odontologo buscar(Integer id) {
-        return odontologoRepository.findById(id).get();
+        Odontologo odontologo = null;
+        Optional<Odontologo> optionalOdontologo = odontologoRepository.findById(id);
+        if (optionalOdontologo.isPresent()) {
+            odontologo = optionalOdontologo.get();
+        }
+        return odontologo;
     }
 
     public List<Odontologo> buscarTodos() {
         return odontologoRepository.findAll();
     }
 
-    public void eliminar(Integer id) {
-        odontologoRepository.deleteById(id);
+    public void eliminar(Integer id) throws ResourceNotFoundException {
+        //odontologoRepository.deleteById(id);
+        if(this.buscar(id) == null){
+            throw new ResourceNotFoundException("Odontologo no existe con id " + id);
+        } else
+        {
+            odontologoRepository.deleteById(id);
+        }
+
     }
 
     public Odontologo actualizar(Odontologo o) {
