+++ src/main/java/clinica/proyectoclinicaodontologica/service/TurnoService.java
@@ -2,12 +2,15 @@ package clinica.proyectoclinicaodontologica.service;
 
 
 
+import clinica.proyectoclinicaodontologica.exceptions.ResourceNotFoundException;
 import clinica.proyectoclinicaodontologica.model.Turno;
 import clinica.proyectoclinicaodontologica.repository.TurnoRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TurnoService {
@@ -26,15 +29,24 @@ public class TurnoService {
     }
 
     public Turno buscar(Integer id) {
-        return turnoRepository.findById(id).get();
+        Turno turno = null;
+        Optional<Turno> optionalTurno = turnoRepository.findById(id);
+        if (optionalTurno.isPresent()) {
+            turno = optionalTurno.get();
+        }
+        return turno;
     }
 
     public List<Turno> buscarTodos() {
         return turnoRepository.findAll();
     }
 
-    public void eliminar(Integer id) {
-        turnoRepository.deleteById(id);
+    public void eliminar(Integer id)  throws ResourceNotFoundException {
+        if (this.buscar(id) == null) {
+            throw new ResourceNotFoundException("turno no existe con id " + id);
+        } else {
+            turnoRepository.deleteById(id);
+        }
     }
 
     public Turno actualizar(Turno t) {
